requirements.txt
discord.py==2.3.2
PyAutoGUI==0.9.54
python-dotenv==1.0.1
Requests==2.31.0
